requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
pyaml
mlflow
dagshub
fastapi
uvicorn 
##-e .